-create a local storage to store theme. -re strcuter files. -remove unused code.
