+++ index.html
@@ -3,12 +3,12 @@
 
 <head>
   <meta charset="UTF-8" />
-  <link rel="icon" type="image/svg+xml" href="/public/favicon.ico" />
+  <link rel="icon" type="image/svg+xml" href="/favicon.ico" />
   <meta name="viewport" content="width=device-width, initial-scale=1.0" />
   <meta name="theme-color" content="#000000" />
   <meta name="description"
     content="TextNinja is used to analyse text and give you access to manipulate text.Whether you want to convert Uppercase to Lowercase or Lowercase to uppercase, reverse text, convert text into title case or perform other transformations like speak the text and copy the text, textNinja has you covered." />
-  <link rel="apple-touch-icon" href="/public/android-icon-192x192.png" />
+  <link rel="apple-touch-icon" href="/android-icon-192x192.png" />
   <title>TextNinja - word counter | character counter | convert to uppercase | convert to lowercase | convert to
     titlecase</title>
   <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
