+++ src/App.jsx
@@ -1,90 +1,26 @@
 import { useState } from "react";
-import Navbar from "./components/Navbar";
-import TextForm from "./components/TextForm";
-import Alert from "./components/Alert";
-import About from "./components/About";
+import TextForm from "./TextForm.jsx";
+import About from "./About.jsx";
 import {
-  BrowserRouter,
-  Routes, // instead of "Switch"
-  Route,
+	BrowserRouter,
+	Routes, // instead of "Switch"
+	Route,
 } from "react-router-dom";
+import { ThemeProvider } from "./context/ThemeContext.tsx";
 
 function App() {
-  const [mode, setMode] = useState("light"); //check dark mode
-  const [alert, setAlert] = useState(null);
-
-  //   const removeBodyClasses=()=>{
-  //   document.body.classList.remove('bg-light')
-  //   document.body.classList.remove('bg-dark')
-  //   document.body.classList.remove('bg-primary')
-  //   document.body.classList.remove('bg-success')
-  //   document.body.classList.remove('bg-warning')
-  //   document.body.classList.remove('bg-danger')
-  // }
-  const showAlert = (message, type) => {
-    setAlert({
-      msg: message,
-      type: type,
-    });
-    setTimeout(() => {
-      setAlert(null);
-    }, 3000);
-  };
-
-  const toggleMode = (cls) => {
-    // removeBodyClasses();
-    // document.body.classList.add('bg-'+cls);
-    // console.log(cls);
-    if (mode === "light") {
-      setMode("dark");
-      document.body.style.backgroundColor = "#0e1a2d";
-      // showAlert("Dark mode has been enabled", "success");
-    } else {
-      setMode("light");
-      document.body.style.backgroundColor = "white";
-      // showAlert("Light mode has been enabled", "success");
-    }
-  };
-  return (
-    <>
-      <BrowserRouter>
-        <Navbar
-          title="TextNinja"
-          about="About us"
-          mode={mode}
-          toggleMode={toggleMode}
-        />
-        <Alert alert={alert} />
-        <div className="container">
-          <Routes>
-            <Route
-              exact
-              path="/"
-              element={
-                <TextForm
-                  heading="Just type in TextNinja - Word Counter, Character Counter, Remove Extra Spaces"
-                  mode={mode}
-                  text="Example Textarea"
-                  showAlert={showAlert}
-                />
-              }
-            />
-          </Routes>
-          <Routes>
-            <Route exact path="/about" element={<About mode={mode} />} />
-          </Routes>
-        </div>
-      </BrowserRouter>
-    </>
-  );
+	return (
+		<ThemeProvider>
+			<BrowserRouter>
+				<Routes>
+					<Route exact path="/" element={<TextForm />} />
+				</Routes>
+				<Routes>
+					<Route exact path="/about" element={<About />} />
+				</Routes>
+			</BrowserRouter>
+		</ThemeProvider>
+	);
 }
-
-// {
-//     <BrowserRouter>
-//       <Routes>
-//         <Route path="/" element={<Home />} />
-//       </Routes>
-//     </BrowserRouter>
-// }
 
 export default App;
